git-svn-id: svn://123.57.173.36/jhj-parent2.0/Branches/jhj-parent2.1@368 57d5540f-7636-4a73-8af3-7dc9f519f443

+++ jhj-dao/src/main/java/com/jhj/po/model/university/PartnerServiceType.java
@@ -27,7 +27,7 @@ public class PartnerServiceType {
 
     private Short serviceProperty;
 
-    private Long serviceTimes;
+    private Double serviceTimes;
 
     private String serviceContent;
 
@@ -127,11 +127,11 @@ public class PartnerServiceType {
         this.serviceProperty = serviceProperty;
     }
 
-    public Long getServiceTimes() {
+    public Double getServiceTimes() {
         return serviceTimes;
     }
 
-    public void setServiceTimes(Long serviceTimes) {
+    public void setServiceTimes(Double serviceTimes) {
         this.serviceTimes = serviceTimes;
     }
